Split pack descriptor into separate header
