+++ src/interface/update.cpp
@@ -18,8 +18,10 @@
 #include "config.hpp"
 #include "coordinates/coordinates.hpp"
 #include "globals.hpp"
+#include "interface/make_pack_descriptor.hpp"
 #include "interface/meshblock_data.hpp"
 #include "interface/metadata.hpp"
+#include "interface/sparse_pack.hpp"
 #include "interface/variable_pack.hpp"
 #include "mesh/mesh.hpp"
 #include "mesh/meshblock.hpp"
